Fix: 프로필 사진 수정 로직 수정 (#270)
* fix: 닉네임 중복 검사 로직 변경 & 만료시간 추가

* fix : 프로필 사진 로직 수정

* fix : 프로필 사진 로직 수정

+++ src/main/java/com/example/dgbackend/domain/member/service/MemberCommandServiceImpl.java
@@ -11,6 +11,7 @@ import com.example.dgbackend.global.exception.ApiException;
 import com.example.dgbackend.global.s3.S3Service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,6 +20,7 @@ import org.springframework.web.multipart.MultipartFile;
 @Service
 @Transactional
 @RequiredArgsConstructor
+@Slf4j
 public class MemberCommandServiceImpl implements MemberCommandService {
 
     @Autowired
@@ -50,7 +52,6 @@ public class MemberCommandServiceImpl implements MemberCommandService {
                 throw new ApiException(ErrorStatus._DUPLICATE_NICKNAME);
             }
             member.setNickName(patchMember.getNickName());
-
         }
 
         member.setName(patchMember.getName());
@@ -66,11 +67,16 @@ public class MemberCommandServiceImpl implements MemberCommandService {
     public MemberResponse.GetMember patchProfileImage(Member member, MultipartFile multipartFile) {
         String originUrl = member.getProfileImageUrl();
 
-        if (originUrl != null) {
+        if (originUrl != null && !originUrl.equals("")) {
             s3Service.deleteFile(originUrl);
         }
-        String profileImageUrl = (s3Service.uploadOneFile(multipartFile).getImgUrl());
-        member.updateProfileImageUrl(profileImageUrl);
+
+        if (multipartFile !=  null) {
+            String profileImageUrl = (s3Service.uploadOneFile(multipartFile).getImgUrl());
+            member.updateProfileImageUrl(profileImageUrl);
+        } else {
+            member.updateProfileImageUrl(null);
+        }
 
         return MemberResponse.toGetMember(member);
     }
